moved theme directory to styles
