+++ src/ui-kit/UserHeader.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import PropTypes from 'prop-types'
 //components
 import Avatar from 'ui-kit/Avatar'
-import {muiColors} from 'lib/theme/colors'
+import {muiColors} from 'styles/theme/colors'
 
 const OuterContainer = styled.div`
   display: flex;
